Refactor donation creation process components
