+++ frontend/src/donation_frontend/components/DonationCreationProcess/DonationCreationForm.svelte
@@ -1,9 +1,9 @@
 <script>
-	import BitcoinStep from './BitcoinStep.svelte';
-	import TransactionStep from './TransactionStep.svelte';
-	import RecipientStep from './RecipientStep.svelte';
-	import DonationStep from './DonationStep.svelte';
-	import ConfirmationStep from './ConfirmationStep.svelte';
+	import CurrencyStep from './CurrencyInfo/CurrencyStep.svelte';
+	import TransactionStep from './Transaction/TransactionStep.svelte';
+	import RecipientStep from './Recipient/RecipientStep.svelte';
+	import DonationStep from './Donation/DonationStep.svelte';
+	import ConfirmationStep from './Confirmation/ConfirmationStep.svelte';
 
 	export let active_step;
 	
@@ -14,12 +14,12 @@
 </script>
 
 <form class="bg-white dark:bg-gray-800 rounded-lg shadow-lg p-12 text-center max-w-full mx-auto" on:submit={handleSubmit}>
-	{#if active_step == 'Bitcoin'}
-		<BitcoinStep />
+	{#if active_step == 'Recipient'}
+		<RecipientStep />
+	{:else if active_step == 'Currency'}
+		<CurrencyStep />
 	{:else if active_step == 'Transaction'}
 		<TransactionStep />
-	{:else if active_step == 'Recipient'}
-		<RecipientStep />
     {:else if active_step == 'Donation'}
   		<DonationStep />
 	{:else if active_step == 'Confirmation'}
